Transferring components to different files to simplify navigation in the project
