+++ Mario/TileTypes.h
@@ -1,0 +1,13 @@
+#pragma once
+
+enum class ETileTextureType
+{
+    Empty = 0,
+    Ground = 1
+};
+
+enum class ETileCollisionType
+{
+    NoCollision,
+    Collision
+};
